requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
pytorch-lightning==0.9.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchmetrics==0.5.0
tensorboard==2.6.0
itk-io==5.2.0.post2
itk-filtering==5.2.0.post2
monai-weekly==0.7.dev2132
nibabel==3.2.1
numpy==1.19.5
opencv-python-headless
Pillow==8.3.1
matplotlib==3.3.4
